Be/feat/nutrient#11

See merge request s09-bigdata-recom-sub2/S09P22C106!29

+++ backend/Foody/src/main/java/com/foody/member/dto/request/MemberInfoModifyRequest.java
@@ -2,8 +2,8 @@ package com.foody.member.dto.request;
 
 public record MemberInfoModifyRequest(
     String nickname,
-    float height,
-    float weight,
+    double height,
+    double weight,
     int gender,
     int age,
     int activityLevel
